ai-documentation-generator/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
openai
gitpython

